refactor(chat): extract the turn engine; settle replies at the token level
commands/chat.rs (2,300 lines) splits: the REPL — input parsing, slash
commands, ghost suggester, conversation state — stays; producing a turn
moves to chat/engine.rs (Turn = Execution × Thinking, TurnRunner, the
tool loop, forced replies, all protocol emission).

The settled reply/thought derivation drops its three string splitters
(split_thought / split_reasoning / strip_thought_open) for the shared
token-level walk: ChannelIds::settle_prethink and ::settle_tool_reply,
with the no-close fallbacks (salvage the answer from the thought's
tail; keep unparsed tool calls whole) preserved and unit-tested against
sentinel channel ids. Prompts are now built against the session
tokenizer instead of reloading tokenizer.json from disk every turn.
