+++ src/chat/mod.rs
@@ -16,6 +16,7 @@
 //!   a `ChatEvent` consumer.
 
 mod decode;
+pub(crate) mod engine;
 mod event;
 mod render;
 mod sink;
@@ -25,4 +26,4 @@ pub use event::ChatEvent;
 pub use render::{ChatStream, StreamDisplay, render_demo};
 pub use sink::{JsonlSink, PlainSink, SharedSinks, SinkSet, emit};
 
-pub(crate) use decode::{ChannelIds, Stabilizer, strip_thought_open};
+pub(crate) use decode::{ChannelIds, Stabilizer};
